Resize HImage's internal _image.
The internal _image (of type PImage) must be resized because certain HYPE funcationality depends on this internal image. An example is HPixelColorist which uses PImage's  function to retrieve the pixel color information. Therefore, without resizing _image, if you pass an HImage that has been resized into HPixelColorist, the pixel colors are retrieved from the original size and not the new size.